Refactor: split networks related functions out of util.py into networks.py
